+++ NEWS.md
@@ -1,5 +1,9 @@
 # polylabelr 0.1.0.9000
 
+## Bug fixes
+
+* The `CITATION` file now correctly displays the year of the package.
+
 # polylabelr 0.1.0
 
 * First release.