+++ NEWS.md
@@ -1,3 +1,5 @@
+# polylabelr 0.2.0
+
 # polylabelr 0.1.0.9000
 
 ## New features